Added Pump Speed Calibration

+++ Scale.h
@@ -5,8 +5,8 @@ class Scale {
 
   Scale(byte, byte);
   HX711 hx711;
-  int value;
-  int previous_value;
+  float value;
+  float previous_value;
   int cal_fact = 370;
 
   void init();
